pyproject.toml
[tool.poetry]
name = "wisdomdrops"
version = "0.1.0"
description = "Using LLMs to distribute knowledge in a structure way"
authors = ["Carlos Morales <camorales@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
langchain = "^0.2.0"
langsmith = "^0.1.60"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
